# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Api
 * Telegram Bot Studio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BotCredentials {
  geminiKey: string;
  botToken: string;
}

export interface BotActionResult {
  success: boolean;
  message: string;
}

export interface ErrorResult {
  error: string;
}

export type LogEntryLevel = typeof LogEntryLevel[keyof typeof LogEntryLevel];


export const LogEntryLevel = {
  info: 'info',
  success: 'success',
  error: 'error',
  warning: 'warning',
  processing: 'processing',
} as const;

export interface LogEntry {
  id: string;
  message: string;
  level: LogEntryLevel;
  time: string;
}

export interface BotStatus {
  running: boolean;
  botName?: string;
  botUsername?: string;
  processedCount: number;
  logs: LogEntry[];
  uptime?: number;
}

export interface TestBotRequest {
  botToken: string;
}

export interface TestBotResult {
  success: boolean;
  botName?: string;
  botUsername?: string;
  error?: string;
}

export type AppSettingsFacebookPublishMethod = typeof AppSettingsFacebookPublishMethod[keyof typeof AppSettingsFacebookPublishMethod];


export const AppSettingsFacebookPublishMethod = {
  token: 'token',
  make: 'make',
  zapier: 'zapier',
} as const;

export interface AppSettings {
  font: string;
  fontSize: number;
  yPosition: number;
  lineHeight: number;
  strokeThickness: number;
  textColor: string;
  activeColor: string;
  ttsSpeed: boolean;
  ttsVoice: string;
  duaaStyle: string;
  videoQuality?: string;
  bgOpacity?: number;
  showBackground?: boolean;
  geminiModel?: string;
  originalVolume?: number;
  duaaVolume?: number;
  wordEffect?: string;
  transitionEffect?: string;
  transitionDuration?: number;
  youtubeToken?: string;
  youtubeClientId?: string;
  youtubeClientSecret?: string;
  tiktokToken?: string;
  facebookToken?: string;
  facebookPageToken?: string;
  facebookPublishEnabled?: boolean;
  makeWebhookUrl?: string;
  zapierWebhookUrl?: string;
  facebookPublishMethod?: AppSettingsFacebookPublishMethod;
  publishDescription?: string;
  aspectRatio?: string;
  scheduledFbPostEnabled?: boolean;
  scheduledFbPostTime?: string;
  scheduledFbPostDays?: string;
  scheduledDuaaStyle?: string;
  youtubeAutoCaption?: boolean;
  captionTranslateLang?: string;
  autoReportEnabled?: boolean;
  autoReportChatId?: string;
  weeklyReportDay?: number;
  smartBotEnabled?: boolean;
  managedChannelIds?: string;
  smartBotAdminChatId?: string;
  muteOriginal?: boolean;
  muteDuaa?: boolean;
  bgColor?: string;
  bgColorMode?: string;
  shadowColor?: string;
  shadowColorMode?: string;
  [key: string]: unknown;
 }