# lib/api-zod/src/generated/types/appSettings.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Api
 * Telegram Bot Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { AppSettingsFacebookPublishMethod } from './appSettingsFacebookPublishMethod';

export interface AppSettings {
  font: string;
  fontSize: number;
  yPosition: number;
  lineHeight: number;
  strokeThickness: number;
  textColor: string;
  activeColor: string;
  ttsSpeed: boolean;
  ttsVoice: string;
  duaaStyle: string;
  videoQuality?: string;
  bgOpacity?: number;
  showBackground?: boolean;
  geminiModel?: string;
  originalVolume?: number;
  duaaVolume?: number;
  wordEffect?: string;
  transitionEffect?: string;
  transitionDuration?: number;
  youtubeToken?: string;
  youtubeClientId?: string;
  youtubeClientSecret?: string;
  tiktokToken?: string;
  facebookToken?: string;
  facebookPageToken?: string;
  facebookPublishEnabled?: boolean;
  makeWebhookUrl?: string;
  zapierWebhookUrl?: string;
  facebookPublishMethod?: AppSettingsFacebookPublishMethod;
  publishDescription?: string;
  aspectRatio?: string;
  scheduledFbPostEnabled?: boolean;
  scheduledFbPostTime?: string;
  scheduledFbPostDays?: string;
  scheduledDuaaStyle?: string;
  youtubeAutoCaption?: boolean;
  captionTranslateLang?: string;
  autoReportEnabled?: boolean;
  autoReportChatId?: string;
  weeklyReportDay?: number;
  smartBotEnabled?: boolean;
  managedChannelIds?: string;
  smartBotAdminChatId?: string;
  muteOriginal?: boolean;
  muteDuaa?: boolean;
  bgColor?: string;
  bgColorMode?: string;
  shadowColor?: string;
  shadowColorMode?: string;
  [key: string]: unknown;
 }